django_api/requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
pytest
django-pytest
pytest-django
psycopg2-binary
pytest-django
python-dotenv
requests
Redis
celery
django-redis
django_celery_results
django_celery_beat
djangorestframework-camel-case
pyhumps
django-polymorphic
django-debug-toolbar
django-environ